Importing more Oculus related classes and organization
Imported more Oculus tools that were left out in the last import that are important to the completion of my prototype.  Also objects in the hierarchy have been organized so the prototype looks neater.  A simple room has also been made for the user and the drum kit to reside in while the user plays the drums.