Attaching drum sticks to the hands and importing audio
The hands in VR should now have properly positioned drum sticks attached to them which will allow the user to play the drums.  Audio has also been imported for each drum so when a user hits a drum with the drumsticks, the correct audio should play depending on what drum they hit.