Almost done
Game is almost done!  Added the floor tom channel, adjusted the speed so it would go at a decent tempo, and make it so if you follow along with the columns, you are able to produce a beat out of them.  All that is left is to fix one small error.